Add files via upload
HexGUI.py is the GUI implementation of Hex game developed by using Python tkinter.
Python tkinter comes by default with most of the python implementation. If tkinter is not installed on the PC, where this program runs, install tkinter first by pip or conda.
To run the program, go to the directory where HexGUI.py file is located and enter 'python HexGUI.py' (without the apostrophes) to start the program.
The function to evaluate a winner is not implemented currently. The Monte-Carlo search method probably will be used for winner evaluation.